requirements.txt
# MSP Toolkit - Production Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
pydantic>=2.5,<3.0
httpx>=0.25,<1.0
jinja2>=3.1,<4.0
schedule>=1.2,<2.0
cryptography>=41.0,<42.0
pyyaml>=6.0,<7.0
click>=8.1,<9.0
rich>=13.7,<14.0
sqlalchemy>=2.0,<3.0
psutil>=5.9,<6.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
structlog>=24.1,<25.0
keyring>=24.3,<25.0

# MCP integration
mcp>=0.9,<1.0

# Development dependencies (optional)
# Install with: pip install -r requirements-dev.txt
